test(server): share lifecycle test builders
Refs #2855
